storage/requirements.txt
Django==4.2.10
django-bootstrap-v5==1.0.11
django-bootstrap3==23.4
django-bootstrap5==24.3
django-cors-headers==4.4.0
djangorestframework==3.14.0
drf-spectacular==0.27.2
gunicorn==21.2.0
Pillow==10.2.0
uwsgi>=2.0.19,<2.1
psycopg2-binary==2.9.7
